// src/common/enum/shipment-status.enum.ts
export enum ShipmentStatus {
    READY_TO_PICKUP = 'READY_TO_PICKUP',
    READY_TO_PICKUP_AT_BRANCH = 'READY_TO_PICKUP_AT_BRANCH',
    READY_TO_DELIVER = 'READY_TO_DELIVER',
    WAITING_PICKUP = 'WAITING_PICKUP',
    PICKED_UP = 'PICKED_UP',
    IN_TRANSIT = 'IN_TRANSIT',
    ARRIVED_AT_BRANCH = 'ARRIVED_AT_BRANCH',
    AT_BRANCH = 'AT_BRANCH',
    DEPARTED_FROM_BRANCH = 'DEPARTED_FROM_BRANCH',
    DELIVERED = 'DELIVERED',
    ON_THE_WAY = 'ON_THE_WAY',
    ON_THE_WAY_TO_ADDRESS = 'ON_THE_WAY_TO_ADDRESS',
    COMPLETED = 'COMPLETED',
}